pyproject.toml
[project]
name = "my-fridge-image-model"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "opencv-python>=4.11.0.86",
    "tensorflow>=2.19,<2.20",
    "tf-keras==2.19.*",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "ultralytics>=8.3.0",
    "roboflow>=1.2.10",
    "python-dotenv",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }